fix: resolve all code review findings for RCU hot-reload
- Heap-allocate VersionedEntitySet to eliminate dangling pointer risk

- Replace detach()/stop() with join() for safe thread cleanup

- Standardize fuzzy_threshold to f32 across all modules

- Add --watch-interval CLI tests (valid, zero, non-numeric)

- Add file_watcher unit tests (getFileStat, start/join lifecycle)

- Add file_watcher to root.zig test discovery

- Fix formatting (extra blank line, trailing newline)

+++ src/versioned_entity_set.zig
@@ -113,7 +113,7 @@ pub const VersionedEntitySet = struct {
 /// The returned snapshot has ref_count = 1 (the "owner" reference).
 pub fn loadSnapshotFromFile(
     path: []const u8,
-    fuzzy_threshold: f64,
+    fuzzy_threshold: f32,
     version: u32,
     allocator: std.mem.Allocator,
 ) !*EntitySnapshot {
